dom/media/tests/mochitest/test_peerConnection_setRemoteAnswerInHaveRemoteOffer.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="mediaStreamPlayback.js"></script>
  <script type="application/javascript" src="pc.js"></script>
  <script type="application/javascript" src="templates.js"></script>
  <script type="application/javascript" src="turnConfig.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "784519",
    title: "setRemoteDescription (answer) in 'have-remote-offer'"
  });

  var test;
  runNetworkTest(function () {
    test = new PeerConnectionTest();
    test.setMediaConstraints([{audio: true}], [{audio: true}]);
    test.chain.removeAfter("PC_REMOTE_SET_REMOTE_DESCRIPTION");

    test.chain.append([[
      "PC_REMOTE_SET_REMOTE_ANSWER",
      function (test) {
        test.pcLocal._last_offer.type="answer";
        test.pcRemote.setRemoteDescriptionAndFail(test.pcLocal._last_offer,
          function(err) {
            is(err.name, "INVALID_STATE", "Error is INVALID_STATE");
            test.next();
          } );
      }
    ]]);

    test.run();
  });
</script>
</pre>
</body>
</html>
